Fix library build step in build.zig
This is temporary - there is no library proper yet! 'proto' will be 
developed first to get a sense of developing in Zig.

[There is no sign of my main computer working again in the near future. 
It will require a motherboard replacement. This is borrowed, and so is 
my time on it. This may make development much slower and relatively 
difficult. Especially considering this is on Windows.]